pyproject.toml
[project]
name = "tiger-lead-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "sqlalchemy==2.0.23",
    "pydantic==2.4.2",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "email-validator==2.1.0.post1",
    "resend>=1.0.0",
    "aiosmtplib==2.0.2",
    "py3-validate-email>=1.0.5.post2",
    "validate-email>=1.3",
    "pandas==2.1.3",
    "openpyxl==3.1.2",
    "stripe==11.2.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "pypdf>=4.0.0",
    "PyMuPDF>=1.23.0",
    "pyotp==2.9.0",
    "qrcode[pil]==8.2",
    "pywebpush==1.14.1",
]
